+++ NVs.OccupancySensor.CV/IMatConverter.cs
@@ -5,6 +5,6 @@ namespace NVs.OccupancySensor.CV
 {
     public interface IMatConverter 
     {
-            Image<Rgb, float> Convert(Mat input);
+            Image<Rgb,byte> Convert(Mat input);
     }
 }
